+++ xmake/modules/private/cache/build_cache.lua
@@ -172,6 +172,7 @@ function dump_stats()
     local cache_hit_total_time = (_g.cache_hit_total_time or 0)
     local cache_miss_total_time = (_g.cache_miss_total_time or 0)
 
+    -- build cache stats
     print("")
     cprint("${color.success}build cache stats:")
     vprint("cache directory: %s", rootdir())
@@ -179,7 +180,7 @@ function dump_stats()
     print("cache hit: %d", cache_hit_count)
     vprint("cache hit total time: %0.3fs", cache_hit_total_time / 1000.0)
     print("cache miss: %d", cache_miss_count)
-    print("cache miss total time: %0.3fs", cache_miss_total_time / 1000.0)
+    vprint("cache miss total time: %0.3fs", cache_miss_total_time / 1000.0)
     print("new cached files: %d", newfiles_count)
     vprint("remote cache hit: %d", remote_hit_count)
     vprint("remote new cached files: %d", remote_newfiles_count)
@@ -187,6 +188,21 @@ function dump_stats()
     vprint("compile fallback count: %d", compile_fallback_count)
     vprint("compile total time: %0.3fs", compile_total_time / 1000.0)
     print("")
+
+    -- file compile time stats
+    if _g.file2compile_time then
+        cprint("${color.success}file compile time (top3) stats:")
+        local sorted_file2compile_time = {}
+        for key, value in pairs(_g.file2compile_time) do
+            table.insert(sorted_file2compile_time, {key = key, value = value})
+        end
+        table.sort(sorted_file2compile_time, function(a, b) return a.value > b.value end)
+        -- print top 3 files with longest compile time
+        for i = 1, math.min(3, #sorted_file2compile_time) do
+            print("%s: %0.3fs", sorted_file2compile_time[i].key, sorted_file2compile_time[i].value / 1000.0)
+        end
+        print("")
+    end
 end
 
 -- get object file
@@ -317,7 +333,12 @@ function build(program, argv, opt)
             if not cppinfo.errdata or #cppinfo.errdata == 0 then
                 cppinfo.errdata = preprocess_errdata
             end
-            _g.compile_total_time = (_g.compile_total_time or 0) + (os.mclock() - compile_start_time)
+            local file2compile_time = _g.file2compile_time or {}
+            _g.file2compile_time = file2compile_time
+            local source_file = cppinfo.sourcefile
+            local compile_time = os.mclock() - compile_start_time
+            _g.file2compile_time[source_file] = compile_time
+            _g.compile_total_time = (_g.compile_total_time or 0) + compile_time
             if cachekey then
                 local extrainfo
                 if cppinfo.outdata and #cppinfo.outdata ~= 0 then