+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### 新特性
 
 * [#15](https://github.com/TOMO-CAT/xmake/pull/15): 添加 `config.debugdir` 函数并将 build batchjobs 写入到 debugdir 便于排查
+* [#17](https://github.com/TOMO-CAT/xmake/pull/17): 设置 test target 的默认超时时间 `run_timeout` 为 60 秒
+* [#18](https://github.com/TOMO-CAT/xmake/pull/18): 默认展示 brief build cache stats
 
 ### Bugs 修复