+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#17](https://github.com/TOMO-CAT/xmake/pull/17): 设置 test target 的默认超时时间 `run_timeout` 为 60 秒
 * [#18](https://github.com/TOMO-CAT/xmake/pull/18): 默认展示 brief build cache stats
 * [#20](https://github.com/TOMO-CAT/xmake/pull/20): `build.ccache.global_storage` policy 默认打开方便多目录复用 ccache
+* [#21](https://github.com/TOMO-CAT/xmake/pull/21): 打印文件编译时长 top3 统计报告
 
 ### Bugs 修复