+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## master
 
-### 新特性
+### v3.0.2
 
 * [#15](https://github.com/TOMO-CAT/xmake/pull/15): 添加 `config.debugdir` 函数并将 build batchjobs 写入到 debugdir 便于排查
 * [#17](https://github.com/TOMO-CAT/xmake/pull/17): 设置 test target 的默认超时时间 `run_timeout` 为 60 秒